Refactor: Apply ts to ProductDetail page
