--- src/pages/ProductDetail/index.js
@@ -1,3 +1,0 @@
-import ProductDetail from "./ProductDetail";
-
-export default ProductDetail;
