+++ src/pages/ProductDetail/index.ts
@@ -1,0 +1,3 @@
+import ProductDetail from './ProductDetail';
+
+export default ProductDetail;
